refactor(detail-refactor): 디테일 페이지에서 불필요하게 컴포넌트 뎊스를 가지는 부분 수정
- 명세에 맞추어 page는 여러 독립적인 section으로 이루어지도록 구성하기 위해서 불필요하게 section 두개를 묶은 부분(ProgramHeader, ProgramDetail)을 따로 독립적인 section으로 수정
- 분리함에 따라서 기존에 두 컴포넌트를 묶어서 로더를 적용하는 부분 분리
