+++ FE/src/app/(admin)/admin/detail/[programId]/page.tsx
@@ -1,5 +1,6 @@
-import AttendeeInfoContainer from "@/components/programDetail/attendee/AttendeeInfo.container";
-import ProgramInfo from "@/components/programDetail/program/ProgramInfo";
+import AttendeeInfoContainer from "@/components/feature/detail/attendee/AttendeeInfo.container";
+import ProgramHeaderSection from "@/components/feature/detail/program/ProgramHeaderSection";
+import ProgramDetailSection from "@/components/feature/detail/program/ProgramDetailSection";
 
 interface ProgramDetailPageProps {
   params: {
@@ -12,7 +13,10 @@ const ProgramDetailPage = ({ params }: ProgramDetailPageProps) => {
 
   return (
     <div className="mb-16 space-y-16">
-      <ProgramInfo programId={+programId} accessType="admin" />
+      <section className="space-y-8">
+        <ProgramHeaderSection />
+        <ProgramDetailSection />
+      </section>
       <AttendeeInfoContainer programId={+programId} isLoggedIn />
     </div>
   );
