+++ FE/src/app/(private)/(program)/detail/[programId]/page.tsx
@@ -1,6 +1,7 @@
-import AttendeeInfoContainer from "@/components/programDetail/attendee/AttendeeInfo.container";
-import ProgramInfo from "@/components/programDetail/program/ProgramInfo";
-import UserAttendModalContainer from "@/components/programDetail/userAttendModal/UserAttendModal.container";
+import AttendeeInfoContainer from "@/components/feature/detail/attendee/AttendeeInfo.container";
+import ProgramHeaderSection from "@/components/feature/detail/program/ProgramHeaderSection";
+import ProgramDetailSection from "@/components/feature/detail/program/ProgramDetailSection";
+import UserAttendModalContainer from "@/components/feature/detail/userAttendModal/UserAttendModal.container";
 
 interface ProgramDetailPageProps {
   params: {
@@ -13,7 +14,10 @@ const ProgramDetailPage = ({ params }: ProgramDetailPageProps) => {
 
   return (
     <div className="mb-16 space-y-16">
-      <ProgramInfo programId={+programId} accessType="private" />
+      <section className="space-y-8">
+        <ProgramHeaderSection />
+        <ProgramDetailSection />
+      </section>
       <AttendeeInfoContainer programId={+programId} isLoggedIn />
       <UserAttendModalContainer programId={+programId} isLoggedIn />
     </div>
