+++ FE/src/components/feature/detail/attendee/AttendeeInfo.container.tsx
