+++ FE/src/components/feature/detail/attendee/AttendeeInfo.loader.tsx
