+++ FE/src/components/feature/detail/attendee/AttendeeInfo.tsx
