+++ FE/src/components/feature/detail/attendee/AttendeeStatus.tsx
