+++ FE/src/components/feature/detail/attendee/BluredAttedee.tsx
