--- FE/src/components/feature/detail/loader/ProgramInfo.loader.tsx
@@ -1,12 +1,0 @@
-import ProgramDetailSkeleton from "./ProgramDetail.skeleton";
-import ProgramHeaderSkeleton from "./ProgramHeader.skeleton";
-
-const ProgramInfoLoader = () => {
-  return (
-    <div className="space-y-8">
-      <ProgramHeaderSkeleton />
-      <ProgramDetailSkeleton />
-    </div>
-  );
-};
-export default ProgramInfoLoader;
