+++ FE/src/components/feature/detail/userAttendModal/AttendStatusView.tsx
