+++ FE/src/hooks/usePrograms.ts
@@ -1,0 +1,6 @@
+import { useParams } from "next/navigation";
+
+export const useGetProgramId = () => {
+  const { programId } = useParams();
+  return +programId;
+};
